Add files via upload
This is a working prototype of VoiBox.net I have included screen shots from mysql in order to recreate the mysql connection in windows. I have also added all of the files from the standalone folder in my wildfly. Remeber that standalone-full.xml is the important one to get right. Even though I have provided most everything I would need to run this off windows you will have to create the data connection and name yourself in the wildfly console, than replace the jndi name if it does not match